Add files via upload
I added some more locations, first names, and surnames.
I commented my code better to say which parts of my code are actually functional yet, or only a work-in-progress.
I defined a clearScreen() function that prints a lot of new lines to de-clutter the terminal between menus. I ask for your patience until I make a GUI.
The menu options look better.
I added something to the user's input; when asked an integer for the amount of friends in the user's party, there is now a loop that verifies that the user's input is indeed an integer. If the user wrote "five" the program couldn't work with it. Maybe later I will consider my users who spell their numbers out, and I'll implement some code to accommodate them. 
I am not sure what I want to do about locations. I have names of locations, and I am not sure whether I want to do an (X,Y) coordinate system, or rather have something like "You are at the library. Your travel options are: the beach, the shop, the cave."
I also want to implement a file save/load system. This shouldn't be difficult. I imagine that when the game launches, a save file .csv existence will be checked. If it exists, it will load the player's party's statistics (Names, levels, money, etc.) If it doesn't exist, the user can save their game progress to a .csv file at any time. I think I will keep the save files in a subfolder of the same folder that the game's .py file is in. The possibilities are endless, I can allow the user multiple saves by allowing them to choose their save file's name and load from a list. Nii