Add files via upload
I have commented out my non-working code that is ambitious. I have a new file locations.py. There is a new Location class, so each location in the game will have a name, a list each of surnames and forenames for that location, and the location's list of possible languages. I want to choose from these randomly, so when NPCs generate, they will have a name and language matching their hometown. I didn't get this working today, so I will think and come back to it.